Fix rounding logic, display rules, update README
Now every number greater than abs(100000.nm) with n,m in [0,1,2,...,9]
(that is the max number of digits displayable) OR smaller than abs(0.1)
is shown in scientific notation with 4 decimals.
IF a number doesn't follow in that case BUT it still has decimals,
those get rounded to the first two digits.